+++ tests/modeling/test_trial_design.py
@@ -1,4 +1,4 @@
-from pharmpy.modeling import add_arm, add_observations, create_trial_design
+from pharmpy.modeling import add_administration, add_arm, add_observations, create_trial_design
 
 
 def test_create_trial_design():
@@ -18,3 +18,10 @@ def test_add_observations():
     td = add_observations(td, arm="Drug", variable="DV", time_points=[0.0, 1.0, 2.0, 4.0, 8.0])
     assert td[0].name == "Drug"
     assert td[0][0].time_points[-1] == 8.0
+
+
+def test_add_administration():
+    td = create_trial_design()
+    td = add_arm(td, name="Drug", size=20)
+    td = add_administration(td, arm="Drug", variable="AMT", amount=10.0, time_points=[0.0, 4.0])
+    assert td[0][0].dose.amount == 10.0